fix: arreglo de subida a cloudinary
Actualizacion de subida de foto de perfil en updateUser

+++ Desarrollo/Artemis/Codificación/BackEnd/dao/user/updateUser.js
@@ -1,6 +1,5 @@
 const User = require("../../models/user")
 const cloudinary = require('cloudinary')
-const underscore = require("underscore")
 const bcrypt = require("bcrypt")
 const fs = require('fs-extra');
 
@@ -15,26 +14,33 @@ let updateUser = async (req, res) => {
 
         if(user==null) { return custom_error_response(400, "Usuario no encontrado")}   
 
-        let body = underscore.pick(req.body, ["nick_name", "artistic_name", "email"])
+        let body = req.body
     
         if (req.body.password) {
             body.password = bcrypt.hashSync(req.body.password, 10)
         }
 
-        const result = await cloudinary.v2.uploader.upload(req.file.path)
-        console.log('Subido a cloudinary')
+        if(req.file != undefined)
+        {
+            const result = await cloudinary.v2.uploader.upload(req.file.path)
+            console.log('Subido a cloudinary')
 
-        //Logica para borrar imagen de cloudinary si el user.img_public_id existe
-        if(user.img_public_id != null) { await cloudinary.v2.uploader.destroy(user.img_public_id)}
-        
-        //cambiar parametros
-        user.profile_img_url = result.url
-        user.img_public_id = result.public_id
+            //Logica para borrar imagen de cloudinary si el user.img_public_id existe
+            if(user.img_public_id != null) { await cloudinary.v2.uploader.destroy(user.img_public_id)}
+            
+            //cambiar parametros
+            user.profile_img_url = result.url
+            user.img_public_id = result.public_id
+        }
+
+        if(body.nick_name){ user.nick_name = body.nick_name }
+        if(body.artistic_name){ user.artistic_name = body.artistic_name }
+        if(body.email){ user.email = body.email }
 
         user.save(async (err)=>{
             if (err) { return error_response(400, res, err)}
 
-            await fs.unlink(req.file.path)
+            if(req.file != undefined) { await fs.unlink(req.file.path) }    
 
             custom_response(res, "Usuario actualizado con éxito")
         })
